chore: setup and correct config files for Pages Router, attempt to make an architecture
