+++ search-app/public/favicon.ico
