--- search-app/src/app/page.tsx
@@ -1,7 +1,0 @@
-// import Image from "next/image";
-
-function Home() {
-  return <h1>Home Page</h1>;
-}
-
-export default Home;
