--- search-app/src/app/search/page.tsx
@@ -1,3 +1,0 @@
-export default function search() {
-  return <div>search</div>;
-}
