+++ search-app/src/pages/App.tsx
@@ -1,14 +1,14 @@
 import "./App.css";
-import { Header } from "./components/header/header";
-import { Footer } from "./components/footer/footer";
+import { Header } from "../components/header/header";
+import { Footer } from "../components/footer/footer";
 import { ReactElement } from "react";
 import { Route, Routes } from "react-router-dom";
 import { ErrorPage } from "./errorPage";
-import { SideSection } from "./components/main/sideSection/sideSection";
-import { MainContent } from "./components/main/mainContent";
-import { ErrorBoundary } from "./components/errorBoundary/errorBoundary";
+import { SideSection } from "../components/main/sideSection/sideSection";
+import { MainContent } from "../components/main/mainContent";
+import { ErrorBoundary } from "../components/errorBoundary/errorBoundary";
 
-export const App = (): ReactElement => {
+const App = (): ReactElement => {
   return (
     <>
       <Header />
@@ -24,3 +24,5 @@ export const App = (): ReactElement => {
     </>
   );
 };
+
+export default App;
