+++ search-app/src/pages/[slug]/client.tsx
@@ -1,0 +1,10 @@
+"use client";
+
+// import React from "react";
+// import dynamic from "next/dynamic";
+
+// const App = dynamic(() => import("../App"), { ssr: false });
+
+export function ClientOnly() {
+  // return <App />;
+}
