+++ search-app/src/pages/errorPage.tsx
