+++ search-app/src/pages/layout.tsx
