+++ search-app/src/pages/page.test.tsx
@@ -1,8 +1,8 @@
+import App from "./App";
 import { render, screen } from "@testing-library/react";
 import { BrowserRouter } from "react-router-dom";
 import { Provider } from "react-redux";
-import { App } from "./App";
-import { store } from "./store/store";
+import { store } from "../store/store";
 
 describe("App", () => {
   test("renders App component", () => {
