+++ search-app/src/pages/page.tsx
@@ -1,0 +1,20 @@
+// import Image from "next/image";
+
+import App from "./App";
+import { Provider } from "react-redux";
+import { BrowserRouter } from "react-router-dom";
+import { store } from "../store/store";
+
+function Home() {
+  return (
+    <>
+      <BrowserRouter>
+        <Provider store={store}>
+          <App />
+        </Provider>
+      </BrowserRouter>
+    </>
+  );
+}
+
+export default Home;
